LPS-189271 Use SearchBuilder and Rename Extension from JS to TS
