+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/common/services/liferay/graphql/account-roles/queries/useGetAccountRolesByAccountExternalReferenceCode.ts
@@ -31,7 +31,7 @@ const GET_ACCOUNT_ROLES_BY_ACCOUNT_EXTERNAL_REFERENCE_CODE = gql`
 `;
 
 export function useGetAccountRolesByAccountExternalReferenceCode(
-	externalReferenceCode,
+	externalReferenceCode: string,
 	options = {
 		filter: '',
 		notifyOnNetworkStatusChange: false,
