+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/ActivationStatus/LiferayExperienceCloud/components/SetupLXCModal/components/SetupLXCForm/components/SetupLXCPage/hooks/useSubmitLXCEnvironment.js
@@ -10,6 +10,7 @@
  */
 
 import {useMutation} from '@apollo/client';
+import SearchBuilder from '~/common/core/SearchBuilder';
 import NotificationQueueService from '../../../../../../../../../../../../../src/common/services/actions/notificationAction';
 import {useAppPropertiesContext} from '../../../../../../../../../../../../common/contexts/AppPropertiesContext';
 import {
@@ -51,7 +52,7 @@ export default function useSubmitLXCEnvironment(
 			const {data} = await client.query({
 				query: getLiferayExperienceCloudEnvironments,
 				variables: {
-					filter: `accountKey eq '${project.accountKey}'`,
+					filter: SearchBuilder.eq('accountKey', project.accountKey),
 				},
 			});
 			if (data) {
